+++ tests/functional/cli/test_factory.py
@@ -9,10 +9,17 @@ from pytest import fixture
 from chalice.cli import factory
 from chalice.deploy.deployer import Deployer, DeploymentReporter
 from chalice.config import Config
+from chalice.config import DeployedResources
 from chalice import local
 from chalice.utils import UI
 from chalice import Chalice
 from chalice.logs import LogRetriever
+from chalice.invoke import LambdaInvokeHandler
+
+
+@fixture
+def no_deployed_values():
+    return DeployedResources({'resources': [], 'schema_version': '2.0'})
 
 
 @fixture
@@ -59,6 +66,29 @@ def test_can_create_botocore_session_debug():
 def test_can_create_botocore_session_connection_timeout():
     session = factory.create_botocore_session(connection_timeout=100)
     assert vars(session.get_default_client_config())['connect_timeout'] == 100
+
+
+def test_can_create_botocore_session_read_timeout():
+    session = factory.create_botocore_session(read_timeout=50)
+    assert vars(session.get_default_client_config())['read_timeout'] == 50
+
+
+def test_can_create_botocore_session_max_retries():
+    session = factory.create_botocore_session(max_retries=2)
+    assert vars(
+        session.get_default_client_config())['retries']['max_attempts'] == 2
+
+
+def test_can_create_botocore_session_with_multiple_configs():
+    session = factory.create_botocore_session(
+        connection_timeout=100,
+        read_timeout=50,
+        max_retries=5,
+    )
+    assert vars(session.get_default_client_config())['connect_timeout'] == 100
+    assert vars(session.get_default_client_config())['read_timeout'] == 50
+    assert vars(
+        session.get_default_client_config())['retries']['max_attempts'] == 5
 
 
 def test_can_create_botocore_session_cli_factory(clifactory):
@@ -201,3 +231,53 @@ def test_can_create_log_retriever(clifactory):
     )
     logs = clifactory.create_log_retriever(session, lambda_arn)
     assert isinstance(logs, LogRetriever)
+
+
+def test_can_create_lambda_invoke_handler(clifactory):
+    lambda_arn = (
+        'arn:aws:lambda:us-west-2:1:function:app-dev-foo'
+    )
+    stage = 'dev'
+    deployed_dir = os.path.join(clifactory.project_dir, '.chalice', 'deployed')
+    os.mkdir(deployed_dir)
+    deployed_file = os.path.join(deployed_dir, '%s.json' % stage)
+    with open(deployed_file, 'w') as f:
+        f.write(json.dumps({
+            'resources': [
+                {
+                    'name': 'foobar',
+                    'resource_type': 'lambda_function',
+                    'lambda_arn': lambda_arn,
+                },
+            ], 'schema_version': '2.0'
+        }))
+
+    invoker = clifactory.create_lambda_invoke_handler('foobar', stage)
+    assert isinstance(invoker, LambdaInvokeHandler)
+
+
+def test_does_raise_not_found_error_when_no_function_found(
+        clifactory, no_deployed_values):
+    with pytest.raises(factory.NoSuchFunctionError) as e:
+        clifactory.create_lambda_invoke_handler('function_name', 'stage')
+    assert e.value.name == 'function_name'
+
+
+def test_does_raise_not_found_error_when_resource_is_not_lambda(clifactory):
+    stage = 'dev'
+    deployed_dir = os.path.join(clifactory.project_dir, '.chalice', 'deployed')
+    os.mkdir(deployed_dir)
+    deployed_file = os.path.join(deployed_dir, '%s.json' % stage)
+    with open(deployed_file, 'w') as f:
+        f.write(json.dumps({
+            'resources': [
+                {
+                    'name': 'foobar',
+                    'resource_type': 'iam_role',
+                    'role_arn': 'bazbuz',
+                },
+            ], 'schema_version': '2.0'
+        }))
+    with pytest.raises(factory.NoSuchFunctionError) as e:
+        clifactory.create_lambda_invoke_handler('foobar', stage)
+    assert e.value.name == 'foobar'